zn - refactoring the Eqsearch and pass the test
