+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -14,7 +14,7 @@ import java.util.HashMap;
 
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 import hello.geojson.FeatureCollection;
-import hello.EqSearch;
+import hello.searches.EqSearch;
 
 @Controller
 public class EarthquakesController {
